refactor : rebuild project structure/ make service, controller
